Add support for addon providing their own layoutlib and/or res.
During addon parsing, the SDK Manager will detect whether the addon
has:
- data/layoutlib.jar
- data/res/
- data/fonts/

if the first one is present, then it is used during rendering.
if *both* the 2nd and 3rd ones are present, then addon resources
are used during rendering.

On the GLE side, all that's needed is adding addons to the list
of rendering targets if they have either library or resources (or
both).

Change-Id: Id16925eea2c98b9fbaaa884ac6fd8c1c1c444db2