Cocoa implementation for SwtMenuBar.
Note that in the Android SDK tools, we ship:
- SWT in 32-bit with Carbon only.
- SWT in 64-bit with Cocoa only.

The previous implementation was carbon-only and the
menus were basically not 'macified' when running on
a recent Mac from the command-line. This missing
implementation fixes it.

After experimenting with various implementations of
the original SWT CocoaMenuEnhancer, I finally settled with
this one since it solely uses reflection and does not
import anything from the swt.cocoa namespace. This means
we can easily build this using the makefile which *only*
links with the 32-bit/carbon version of SWT.jar.

Note that on Windows/Linux, the src-darwin folder will
be ignored, which is why it is not mapped as a source
directory and which is why we can't build this directly
from Eclipse.

Change-Id: I53859d3b15bc7026d6bd4f77e048a0c4b4eeb02c