Fix image reousrces in HierarchyViewer application.
Currently the HierarchyViewer application crashes when
trying to run it from Eclipse. The issue is that the
code looks for all the images in /images/ whereas the
source code places them in /resources/images/ and the
makefile was setup to "erase" the /resources part.

The fix is:
- move all images back to hierarchyviewerlib/src/images.
- adjust the makefile accordingly.

Another issue is that the about box JPEG files failed
to load (at least on Eclipse 3.5 under Win7). That's OK
we had a real icon in PNG that works, so we should be
using it for the about box anyway.

Change-Id: Id3edc3d714480d4251b190706371db2481a14b63