Fix space-in-dir issue with Proguard for the ant case.
This fixes the proper way to quote the compound paths
arguments given to proguard. Each individual path is
quoted if it contains spaces.

Quoting for -libraryjars fixes the issue for the SDK
install path containing a space.
Quoting for -injars and the other arguments fixes the
issue for the project path containing a space (although
the ant build will fail later at the dx phase in this
case.)

Change-Id: I74283b4f1b19c353c1a5ed36e4d0ec58676b6a79