Custom View handling improvements
First and foremost, allow custom views to accept children such that
you can drag & drop children into the custom view in the outline.

Second, prevent an NPE which can occur if you drag into a layout where
the root element is a custom view.

Third, handle <view> (not <View>) better: provide a custom icon, and
inline the view class name in the outline label.

Fourth, allow double clicks (in addition to ctrl-click which is already
supported) on the custom views in the palette to allow jumping to the
custom view code.

Change-Id: I13c2bf2f4169185c9fcc893ce487f2abdac46802