Fix Up/Down buttons in UiTreeBlock form editor.
In the manifest editor (and actually all other form
editors), it was possible move elements up/down
using the Up/Down buttons. However this allowed a
user to move an element in a parent that would not
accept that parent. The tree block would then not
display the element any more and a user would have
had to switch to the XML view to correct this.

This fix thus makes sure a parent will accept the
node being moved. The up/down buttons are also
grayed appropriately if the action isn't possible.

SDK Bugs: 2274556, 2274575

Change-Id: If1dd61f1260063e8ecb9c48330e6c6b2dc3c7228