Fragment Rendering Support
This changeset adds designtime-previewing of fragments, where
a layout which contains fragments will show the contents of
the fragments inline.

Initially, the fragments are empty, but you can right click on them to
bring up a context menu where you can choose which layout to show at
designtime. This is persisted across IDE sessions, just like the
ListView render preview.

In addition to the generic layout chooser, all layout references found
in the associated Fragment class (usually what you want) are listed
directly in the menu.

Change-Id: Iff3e6e5cbf87a19bdb61e677a63f663ca6e4848b