Fix bug in editor open utility
The code to open a file (called by Go To Declaration, Show Include
etc) handles two scenarios:

(1) The file is in the workspace - open using Eclipse IFile mechanism
(2) The file is outside the workspace - open using the fallback
    external storage (which means you get a plain XML editor)

There's a third scenario: the file is not in the workspace, but is
part of a project in the workspace so it does have a valid IFile.
(This can happen if you import a project but choose not to copy the
contents into the workspace).  This changeset adjusts the code to open
up an editor such that it handles this third scenario and we get our
own XML editors for these types of files.

Change-Id: I040e1b899cd38bbda3fcf3475cc4dfb541d10016