Mechanism to disable SwtMenuBar in SdkMan2
The SDK Manager 2 should not try to change the
menu bar when invoked from Eclipse. Actually it
can't because the SwtMenuBar lib isn't even present
to it crashes with a NoClassDef exception.

This CL adds the notion of "invocation context" so
that we know what is invoking the manager and we can
change the UI accordingly.

Change-Id: I606850a20fbc5f9d2d1d4fd0e16aa0bd71ef41c7