Improve the choice of default configuration
When you open a new layout, the layout editor needs to pick an initial
configuration for the layout.  Currently, the choice of layout is
based on the project API level (such that it picks tablet screen and
landscape orientation for API 11, and phone and portrait for lower
APIs).

This changeset adds another factor: the currently used configuration.
If you have another layout visible and you open a new layout, then the
configuration for the current layout will be used for the new layout
(if it is a compatible match).

Change-Id: I601b4373514cb30d6051e9de1bfa2ad91ff1b5f4