Migrate render target settings to released versions
When loading the project's render target setting (which as of ADT 11
is per project rather than per layout), check to see if it has a code
name, and if so, see if the user has the final version of the platform
installed as well. If so, migrate the render target to the final
version instead since the user probably did not intentionally want to
stay with the pre-release version after installing the final version.

Change-Id: I4cdfbb9670a042cf342d2c371ef123eff3959ff2