Offer to install fragment compatibility library
This changeset changes the popup displayed when you drop a fragment
tag in a pre-API-11 project.  Instead of just displaying an error
message, you now get a dialog which asks if you want to install the
project, and if you click the "Install" button it invokes the SDK
manager to install the right package, and then copies it into the
project.

In addition, the fragment selector now has a "Create New" action which
invokes the New Class wizard pre-configured with the right fragment
class.

Change-Id: I4ce1b51fbfe939a21e91379eb6b77ff722c13f0f