New Repository XSD for R12.
This add a new schema sdk-repository-4 with the following
changes:

- <extra> element now has a <project-files> element that
contains 1 or or more <path>, each indicating the relative
path of a file that this package can contribute to
installed projects.

- <platform> element now has a mandatory <layoutlib>
that indicates the API and revision of that layout library
for this particular platform.

There's also a matching sdk-addon-2 schema with the same
modification to the <extra> element.

Change-Id: Id74a5d09dfa36d65f11b9d60a82fe8ba4119eb34