Don't apply default attributes in a move/copy
There is a method in the layout editor which fills in required default
attributes like id, width and height, as well as some optional
attributes like a default text attribute.

This code was called unconditionally whenever a new child is added to
a layout. However, this should only be done when a new widget is
created, not during moves or copies from existing elements.

Change-Id: Ic04dec746c70359eca38454bacfdcdb87891719f