Don't apply default attributes in a move/copy. DO NOT MERGE
There is a method in the layout editor which fills in required default
attributes like id, width and height, as well as some optional
attributes like a default text attribute.

This code was called unconditionally whenever a new child is added to
a layout. However, this should only be done when a new widget is
created, not during moves or copies from existing elements.

Change-Id: I2b437ef609a168f1bb1b7698d3a954a4d2f5faa0