Handle fragment tags in included contexts
This changeset fixes the support for rendering user-chosen layouts in
<fragment> tags in a couple of scenarios:
- where the fragment is part of a layout that "surrounds" the current
  layout (e.g. "show included in")
- where the fragment is part of a layout that is included from the
  edited layout

In both cases, the fragment is read from a layout that is parsed with
a plain pull parser rather than the full XML node model used for
edited layouts, so to fix this the ContextPullParser now looks
specially for <fragment> tags and handles these specially.

In the second scenario, the parser is constructed by layoutlib, so we
have to use the project callback to anticipate requested parsers and
supply our own fragment-capable pull parsers instead.

Change-Id: I8aba253ac0c1a0d5dad7bc6efb9def5f20c75144