New plug-in supporting NDK with CDT
Provides the Add Native Support menu item for Android projects.
Provides build integration for ndk-build and Android toolchains.
No debug support yet, but that's in the works.
Includes a new feature for this that brings in CDT at install time.
Scanner Discovery currently doesn't work on Windows but does on Mac/Linux.
Fixed to adhere to Android coding standards.

Change-Id: I26ddd622aaa3256e336804b1ace8163ed742e9b2
Signed-off-by: Doug Schaefer <cdtdoug@gmail.com>