Code completion improvements inside custom views
This CL fixes several issues when using code completion inside a tag
that is a custom view. In that case we don't have a valid descriptor
which specifies the allowable children, which meant completing inside
any of the children, even if they are standard Android widgets, would
not locate names and attributes. This fixes that, along with handling
dots in tag names which caused other code completion problems.

Change-Id: If0f007322c6aa00363ab32baf6191a3d3ffb9883