Disable PostComp in ADT except on Export or Debug
This patch adds a check in PostCompilerBuilder so if a flag is
not explicitly set (as is done by ExportHelper and Launch) then
it will skip and not do packaging. This should speed up the incremental
build on file save at the cost of slightly longer builds when pushing
to the debugger.

Change-Id: I2993f91a9b53e767e77da82de61b2699dd1127a4