Custom sort property sheet categories
The property sheet is hardcoded to sort property categories
alphabetically.

Unfortunately, this means that for most widgets, the first properties
listed in the property sheet is "Deprecated" !

This changeset adds a custom sorter to the propertysheet, and sorts
the deprecated category to the bottom. It also sorts the "Misc"
category second to last, which means that you'll usually see the type
of the widget (e.g. "Button" or "EditText") at the top.

Change-Id: Ib2c8ec81cae82650a8a41e533e3803a1f6582fa7