Improve error handling for missing and broken custom view classes
When a view class cannot be found during rendering, the error log
lists the class as missing, and presents the class as a
hyperlink. Clicking the link will open the New Class wizard.

This CL changes this such that there are multiple links instead:

- "Fix Build Path": This opens the Java Build Path preference page in
  the Options dialog for the current project where you can fix the
  build path of the project, since a more common source of missing
  classes is a problem with the class path.

- "Edit XML": This switches to the XML tab where you can edit the name
  of the custom view tag: A "missing" view is often a typo in the view
  name

- "Create Class": This opens the create class wizard as before.

Similarly, for broken classes (where the class is found, but cannot be
instantiated correctly) the following hyperlinks are offered:

- "Open Class", which shows the class in the editor

- "Show Error Log", which will open the IDE error log where you
  typically can find a full stack trace from the failed invocation.

In addition, this CL fixes
14546: Creating a custom view from the error output of the layout
editor creates it in gen
by ensuring that the gen/ folder is not one of the fragment roots
presented to the New Class wizard.

Change-Id: I06b8969800caaea719e4b52608a7ed130c91837d