Suggest similar class names or missing pkgs in the error console
If the layout XML file contains typos, the rendering will fail and the
canvas will list the missing classes along with hyperlinks to create a
new class, configure the build path etc.

This changeset looks for "typos" in the view names and if it finds a
similar real view class, either among the Android views or among the
custom views in the current project, then it will add a hyperlink
suggestion to fix the XML by editing the name to the correct spelling.

It also handles the scenario where you have typed in a custom view
class name correctly, but have forgotten to include its package.

In a followup changeset this functionality will be available from the
XML editing quick assistant as well.

Change-Id: Iaefd3f503795e25e6eb38353c60c645061d4814e