Add support for tracking cpu and real time.
Enhanced TraceView to support extracting cpu and real time profile
data from traces and plotting them accordingly.

In contrast, the original thread-cpu algorithm incorrectly assumed
that threads could not execute in parallel so it would produce
inaccurate representations of the global execution timeline on SMP
systems.

To use this feature, Dalvik must be restarted with appropriate
profiling options.

Collect CPU time only (default).  Global timeline must be inferred.
  adb shell setprop dalvik.vm.extra-opts -Xprofile:threadcpuclock

Collect real time only.  CPU usage information is not available.
  adb shell setprop dalvik.vm.extra-opts -Xprofile:wallclock

Collect CPU time and real time information.  Most accurate but
requires a newer VM.
  adb shell setprop dalvik.vm.extra-opts -Xprofile:dualclock

TraceView remains compatible with traces gathered on older devices
and produces identical output when analyzing traces gathered with
the thread-cpu clock.

Added a feature that displays the exclusive and inclusive time
information while hovering over a method.

Added a feature that displays the time spent in a context switch
by a simple 1 pixel underline within the body of the method call.
This makes it much easier to identify long running methods that
spend most of their time blocked on a wait.

Fixed a bug in the way that the method call highlights were
displayed which could cause them to be truncated down to a
single-pixel width depending on the pixel alignment of the
start and end of each segment.

Change-Id: Id2e6d101d9c604a0029e0920127871f88013faa0