Sdkman2: Simplify packages names, fix extra package sort.
Simplify package names when sorted by API, that is
don't repeat "API nn" or "Android x.y" in the package
names if it matches to category.

Fix sorting of packages to properly sort extras
by vendor+path. The previous sort was broken as
it was making the revision number more important
than the vendor+path for extras.

Change-Id: I51ed0114ab9c77ef5d94710fc2449e5be9cd2ad9