When using prerelease SDK versions, use codename instead of API level
The New Project wizard now automatically fills in the Minimum SDK text
field based on the selection in the SDK target chooser. However, it
was hardcoded to always pick the API level number. That's not correct
when using SDK versions with a code name; for example, when used with
a dev SDK build the minimum SDK version should be "AOSP".

Change-Id: I28273007be7e56f3a34b683f01234d5d1046aa0a