Add AdapterView handling for Spinners
This CL adds AdapterView handling similar to what we already provide
for ListViews and GridViews for Spinners.  This means that it will
provide an adapterbinding which provides a layout to use at
designtime. It also adds the ability to pick a different layout to use
at designtime.

More importantly, this fixes rendering of Spinners on Theme.Holo.

Change-Id: I30cef98d22ed57b2e698c4590948e25bf3fecedd