Fix resize feedback for cell spanning mode
When resizing in a GridLayout you can toggle whether you are resizing
the widget or the cell bounds with the Shift key (and the resize
feedback tells you this).

However, the drawn resize feedback was always showing the widget
bounds. This CL fixes this such that when you are resizing the cell
bounds, the current cell bounds are displayed instead.

(Also contains a simple fix to ensure feedback is drawn when you
click and hold (without dragging yet) on a resize handle.)

Change-Id: I7cf24f4508aa5bbafd45f36b13be41e66ecbc6cc