Add tooltip drag feedback for relative layout
This changeset adds tooltip drag feedback for RelativeLayout. During
dragging, the matched edge can quickly change from the left side to
the right side or the top to the bottom edge, so the tooltip support
now allows a client to specify where to align the tooltip not just at
creation but also during updates. This is particularly important for
RelativeLayout where we need to ensure that the tooltips do not
obscure the constraint feedback arrows.  And to prevent flicker, it
uses a timer to ensure that there is at least 750 ms between the
repositioning of the tooltip (but it will apply a side change
immediately if it has been 750 ms since the last time the alignment
changed).

Change-Id: I8817db67844ad13c651d343ad310d8646e34dfca