Adjust the match position description to be relative to the layout
Currently the matches displayed for absolute X and Y locations will
list the coordinates relative to the entire screen, not to the
surrounding GridLayout. This CL changes the feedback to be relative to
the GridLayout bounds.

Also fix a bug in updating tooltips/messages which would call tooltip
flicker (because the message update method was called too early,
before the message object was ready, causing the tooltip to be
removed, then reshown.)

Change-Id: I665586dbcb1f3f6cdcb3a3f475cf363cd258bf83