Asset Studio wizard
This is an initial integration of the Android Asset Studio into
Eclipse, as a New Asset wizard. It uses the Java port of the Android
Asset Studio to generate the assets:
  https://code.google.com/a/google.com/p/android-asset-studio-java/
It only supports launcher icons, and some of the configurable
parameters (file, shape, crop).

To run it, put the AssetStudioLib.jar file into the the adt libs
directory.

There's a new "Asset Set" wizard in the New wizard, which will open up
a two page wizard; the first page lets you select the project (which
it attempts to pick up from context), as well as the type of asset to
create, and the asset output name.

In the second page there are the various widgets to tweak the code
generator, and a preview area on the right which updates as you tweak
the various controls.

The main remaining work is to support additional asset types as
they are added to the assetstudio generator library.

Change-Id: I2e556337f8e5c3bc09e84b35a342ba05110abebc