SDK Manager2: revamp package diff logic.
This revamps the diff algorithm used to merge
the local and remote packages found during a
repository load into the PkgItems displayed in
the tree viewer.

FYI all the Package and Archive sub-classes are getting
an equals() and hashCode() methods to make them behave
correctly when put it in a HashTable or Set. These are
the auto-generated stuff from Eclipse.

Change-Id: I1494a0ed44cd768eed252e3a81b9e74bf86d563c