Fix code completion handling for text selection
This CL fixes
18607: ADT 12 problems with xml files

The bug is that if there is a text selection and you invoke code
completion and select an alternative, the wrong number of characters
is deleted as part of the replacement.

The fix is simple: remove the code which adds in the selection count
from the replacement string, since the complete replacement count is
now already computed properly without regard for the selection (we
should -always- replace the complete value token, not just the
selection).

Most of the diffs actually pertain to new unit tests for this because
the testing infrastructure needed some changes to handle references to
selection ranges rather than just caret offsets.

There are also a couple of updates to some test goldenfiles unrelated
to this fix, which updates the golden files to reflect recent changes
like the package name of GestureOverlayViews and the @attr resource
type in completion.

Change-Id: Ie4f680f201930414457a79846f722db6efc89f93