Customize Quick Outline for Layout and Resource editors
The XML editor has a Quick Outline (Cmd-O) but it only displays the
tag name. This means that you only see "Button" instead of ids to
distinguish similar widgets - and it also means that you can't filter
based on the ids or names.

This changeset adds a custom configuration for the XML quick outline
used by the layout editor and the resource editor which includes the
id or name attributes, and it also picks up the palette icons for the
various widgets.

Change-Id: I08e5d8d10f8f2ad0c2633488d3520d48da6c20cb