Add a new XML formatter, new options, and format on save
This changeset adds a custom XML formatter for the Android XML
editors.

There is an option for turning off this formatter and using the
standard XML formatter instead.

The new custom formatter offers the following features:

* By default, it formats files following the Android conventions for
  XML formatting. This means not only that it indents with 4 space
  characters (instead of the Eclipse default of one tab character),
  but it places attributes on a line of their own, it adds blank
  lines in certain places, and so on.

* It sorts attributes (logically, alphabetically, or none)

* It formats files with different styles depending on the context.
  For example, in layout files it always places a single blank line
  between view elements, whereas in manifest files it only places
  blank lines between elements of different types such that for
  example the <uses-permission> elements all form a logical group,
  separated from <uses-sdk> and so on.

  The fact that strings.xml and layout.xml are usually formatted with
  different flavors is one of the reasons setting up the old Eclipse
  XML formatter to do the right thing was not possible.

In addition, there are various user options added to the Editors
preference page where you can tweak the behavior of the formatter,
such as turning off blank lines completely, controlling whether there
is a space before the > or /> closing characters, and you can also
tell the formatter to use Eclipse's indentation setting (indentation
size and whether to use spaces or tabs).

Finally, there is a new "Format on Save" option which behaves just
like Format on Save for Java files: when you press Ctrl-S to save a
file, the XML content will be formatted before the file is saved.

Change-Id: Ic6f05c768ab063b09f6f0248f60fbe8722fb149d