Improve smart-indent handling in XML files
This changeset adds a new "auto edit strategy" for Android XML files,
which replaces the default XML auto-indenter (which just copied the
indentation of the previous line.)

The new indenter instead uses the lexical information to indent based
on the tag and bracket balance.

For example, if the | represents the caret position, then the
following shows how the edits are transformed:

    <foo name="value"/>|       =>       <foo name="value"/>
                                        |

    <foo name="value">|        =>       <foo name="value">
                                            |

    <foo                                <foo
        attr1 = "value1"|>     =>           attr1 = "value1"
                                            |>

    <foo                                <foo
        attr1 = "value1"/>|    =>           attr1 = "value1"/>
                                        |

It can also modify the text after the caret, as in the following
transformation:

    <foo>|</foo>               =>       <foo>
                                            |
                                        </foo>

See the unit test for more cases.

Change-Id: Ia99ecf6573ff4a9473970aa2fd481d2228ddf45d